! 2009-11-10 - Chris Jones  - Update Muon and Rich PID packers to pack DLLs to integers
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@37740 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 72817c2e75c8860a2d38d1c62980a989a45f0acf

+++ LHCb/Event/PackedEvent/Event/PackedMuonPID.h
@@ -1,4 +1,4 @@
-// $Id: PackedMuonPID.h,v 1.2 2009-11-07 12:20:26 jonrob Exp $
+// $Id: PackedMuonPID.h,v 1.3 2009-11-10 10:24:09 jonrob Exp $
 #ifndef EVENT_PACKEDMUONPID_H
 #define EVENT_PACKEDMUONPID_H 1
 
@@ -37,8 +37,8 @@ namespace LHCb
         mutrack(-1)
     {}
 
-    float MuonLLMu;
-    float MuonLLBg;
+    int MuonLLMu;
+    int MuonLLBg;
     int nShared;
     int status;
     int idtrack;
